improve validation
