--- packages/atlas/src/hooks/useCreateEditMember.tsx
@@ -1,85 +1,0 @@
-import { useApolloClient } from '@apollo/client'
-import { zodResolver } from '@hookform/resolvers/zod'
-import debouncePromise from 'awesome-debounce-promise'
-import { useRef } from 'react'
-import { useForm } from 'react-hook-form'
-import { z } from 'zod'
-
-import { GetMembershipDocument, GetMembershipQuery, GetMembershipQueryVariables } from '@/api/queries'
-import { ImageInputFile } from '@/components/_inputs/MultiFileSelect'
-import { MEMBERSHIP_NAME_PATTERN } from '@/config/regex'
-
-export type EditMemberFormInputs = {
-  handle: string | null
-  avatar: ImageInputFile
-  about: string | null
-}
-
-export const useCreateEditMemberForm = (prevHandle?: string) => {
-  const client = useApolloClient()
-
-  const debouncedHandleUniqueValidation = useRef(
-    debouncePromise(async (value: string, prevValue?: string) => {
-      if (prevValue != null && value === prevValue) {
-        return true
-      }
-      const {
-        data: { membershipByUniqueInput },
-      } = await client.query<GetMembershipQuery, GetMembershipQueryVariables>({
-        query: GetMembershipDocument,
-        variables: { where: { handle: value } },
-      })
-      return !membershipByUniqueInput
-    }, 500)
-  )
-
-  const schema = z.object({
-    handle: z
-      .string()
-      .min(1, { message: 'Member handle cannot be empty' })
-      .min(5, {
-        message: 'Member handle must be at least 5 characters',
-      })
-      .max(40, {
-        message: `Member handle cannot be longer than 40 characters`,
-      })
-      .refine((val) => (val ? MEMBERSHIP_NAME_PATTERN.test(val) : true), {
-        message: 'Member handle may contain only lowercase letters, numbers and underscores',
-      })
-      .refine((val) => debouncedHandleUniqueValidation.current(val, prevHandle), {
-        message: 'Member handle already in use',
-      }),
-    avatar: z.any().nullable(),
-    about: z.string().max(1000, { message: 'About cannot be longer than 1000 characters' }).nullable(),
-  })
-
-  const {
-    register,
-    handleSubmit,
-    setFocus,
-    getValues,
-    reset,
-    watch,
-    control,
-    formState: { errors, isDirty, isValid, dirtyFields, isValidating },
-  } = useForm<EditMemberFormInputs>({
-    mode: 'onChange',
-    resolver: zodResolver(schema),
-    shouldFocusError: true,
-  })
-
-  return {
-    register,
-    handleSubmit,
-    getValues,
-    reset,
-    watch,
-    setFocus,
-    control,
-    errors,
-    isDirty,
-    isValid,
-    isValidating,
-    dirtyFields,
-  }
-}
